requirements.txt
asgiref==3.7.2
Django==4.0
psycopg2-binary==2.9.2
sqlparse==0.4.4
pillow==9.3.0
djangorestframework==3.13.1
django-filter==21.1
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
django-cors-headers==3.11.0

